Refactors components and fixes minor issues
Moves components into `includes` and `layout` directories for better organization.

Updates imports to reflect the new file structure.

Fixes a minor styling issue in the Fallmanager component.
