+++ gl-core/components/includes/Status.tsx
@@ -1,8 +1,8 @@
 'use client';
 import * as React from 'react';
 import { Alert, Button, IconButton } from '@mui/material';
-import { useSlice, Icon, useDispatch } from '../../gl-core';
-import { toggleStatus } from '../';
+import { useSlice, Icon, useDispatch } from '../../../gl-core';
+import { toggleStatus } from '../../../gl-core';
 
 export type TStatus = {
   title?: string | null;
