+++ gl-core/components/includes/index.tsx
@@ -1,0 +1,5 @@
+import Advert from './Advert';
+import Feedback from './Feedback';
+import Status from './Status';
+
+export { Advert, Feedback, Status };
