--- gl-core/components/layout/Responsive.tsx
@@ -1,20 +1,0 @@
-'use client';
-import React from 'react';
-// import Image from 'next/image';
-import { Box, Container, Grid } from '@mui/material';
-import { Header } from '../../../gl-core';
-
-export type TResponsive = {
-  body?: string;
-  frontmatter?: any;
-};
-
-export default function Responsive({
-  body = '',
-  frontmatter = null,
-}: TResponsive) {
-  //   const isMobile = useIsMobile();
-  const featuredImage = frontmatter?.image || null;
-
-  return <Container>Flash</Container>;
-}
