+++ gl-core/components/layout/index.tsx
@@ -1,15 +1,5 @@
 import Header from './Header';
-import { THeader } from './Header';
-
 import Footer from './Footer';
-import { TFooter } from './Footer';
-
 import RenderMarkdown from './RenderMarkdown';
-import { TRenderMarkdown } from './RenderMarkdown';
 
-import Responsive from './Responsive';
-import { TResponsive } from './Responsive';
-
-export { Header, Footer, RenderMarkdown, Responsive };
-
-export type { THeader, TFooter, TRenderMarkdown, TResponsive };
+export { Header, Footer, RenderMarkdown };
