+++ gl-core/components/nav/index.tsx
@@ -1,22 +1,13 @@
 import Nav from './Nav';
-import { TNav } from './Nav';
 import NavItem from './NavItem';
-import { TNavItem } from './NavItem';
 import { PageBreadcrumb } from './PageBreadcrumb';
 import ShareThis from './ShareThis';
-import { TShareThis } from './ShareThis';
 import MightyButton from './MightyButton';
 import SideAds from './SideAds';
-import { TSideAds } from './SideAds';
-import { TMightyButton } from './MightyButton';
-import CardButton from './CardButton';
-import { TCardButton } from './CardButton';
 import Search from './Search';
-import { TSearch } from './Search';
 import NextPrevious from './NextPrevious';
 import TopRightMenu from './TopRightMenu';
 import ShareMenu from './ShareMenu';
-import { TShareMenu } from './ShareMenu';
 export {
   NextPrevious,
   Nav,
@@ -24,20 +15,8 @@ export {
   PageBreadcrumb,
   ShareThis,
   MightyButton,
-  CardButton,
   Search,
   SideAds,
   TopRightMenu,
   ShareMenu,
-};
-
-export type {
-  TSideAds,
-  TCardButton,
-  TSearch,
-  TMightyButton,
-  TNav,
-  TNavItem,
-  TShareThis,
-  TShareMenu,
 };
